mobile/requirements.txt
# Mobile dependencies
kivy==2.3.0
kivymd==1.2.0
requests==2.31.0
python-dotenv==1.0.1
plyer==2.1.0            # Push notifications, camera, etc.
